+++ NEWS.md
@@ -3,9 +3,11 @@
 
 ## v0.3.0
 
+- Better log formatting
 - `speculate`
     - Use `compile = true` instead of `dry = false` to compile generated signatures
 - `Verbosity`
+    - `debug` has been split into `compile` and `pass`
     - Is now a subtype of `AbstractSet{Verbosity}`
     - Now implements the iteration interface
     - Implement `instances(::Type{Verbosity})`